+++ src/POEMS/fix_poems.cpp
@@ -1525,9 +1525,9 @@ void FixPOEMS::copy_arrays(int i, int j, int /* delflag */)
 double FixPOEMS::memory_usage()
 {
   int nmax = atom->nmax;
-  double bytes = nmax * sizeof(int);
-  bytes += nmax*MAXBODY * sizeof(int);
-  bytes += nmax*3 * sizeof(double);
+  double bytes = (double)nmax * sizeof(int);
+  bytes += (double)nmax*MAXBODY * sizeof(int);
+  bytes += (double)nmax*3 * sizeof(double);
   return bytes;
 }
