+++ src/USER-AWPMD/pair_awpmd_cut.cpp
@@ -734,8 +734,8 @@ void PairAWPMDCut::min_x_set(int /* ignore */)
 
 double PairAWPMDCut::memory_usage()
 {
-  double bytes = maxeatom * sizeof(double);
-  bytes += maxvatom*6 * sizeof(double);
-  bytes += 2 * nmax * sizeof(double);
+  double bytes = (double)maxeatom * sizeof(double);
+  bytes += (double)maxvatom*6 * sizeof(double);
+  bytes += (double)2 * nmax * sizeof(double);
   return bytes;
 }
