+++ src/USER-COLVARS/fix_colvars.cpp
@@ -1000,6 +1000,6 @@ double FixColvars::compute_scalar()
 double FixColvars::memory_usage(void)
 {
   double bytes = (double) (num_coords * (2*sizeof(int)+3*sizeof(double)));
-  bytes += (double) (nmax*size_one) + sizeof(this);
+  bytes += (double)(double) (nmax*size_one) + sizeof(this);
   return bytes;
 }
