+++ src/USER-FEP/pair_lj_cut_tip4p_long_soft.cpp
@@ -588,8 +588,8 @@ void *PairLJCutTIP4PLongSoft::extract(const char *str, int &dim)
 
 double PairLJCutTIP4PLongSoft::memory_usage()
 {
-  double bytes = maxeatom * sizeof(double);
-  bytes += maxvatom*6 * sizeof(double);
-  bytes += 2 * nmax * sizeof(double);
+  double bytes = (double)maxeatom * sizeof(double);
+  bytes += (double)maxvatom*6 * sizeof(double);
+  bytes += (double)2 * nmax * sizeof(double);
   return bytes;
 }
