+++ src/USER-MISC/compute_cnp_atom.cpp
@@ -323,8 +323,8 @@ void ComputeCNPAtom::compute_peratom()
 
 double ComputeCNPAtom::memory_usage()
 {
-  double bytes = nmax * sizeof(int);
-  bytes += nmax * MAXNEAR * sizeof(int);
-  bytes += nmax * sizeof(double);
+  double bytes = (double)nmax * sizeof(int);
+  bytes += (double)nmax * MAXNEAR * sizeof(int);
+  bytes += (double)nmax * sizeof(double);
   return bytes;
 }
